quality control for all 3 (renamed scripts)
